medizap-chatbot-api/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6

# Data handling and numerical operations
pandas==2.1.0
numpy==1.26.4

# Google Cloud and Firebase
firebase-admin==6.4.0
google-cloud-vision==3.10.2

# Machine Learning
scikit-learn==1.3.0
joblib==1.4.2

# API requests
requests==2.31.0

# NEWLY ADDED for Local LLM and Image Handling
llama-cpp-python
Pillow
